Patch request

+++ backend/api/views/cart_view.py
@@ -38,22 +38,10 @@ def verify_quantity(quantity):
     """verify quantity"""
     if type(quantity) != int:
         abort(400, description="quantity value must be an integer")
+    if int(quantity) < 0:
+        abort(400, description="quantity value must be positive")
+
     return quantity
-
-
-def validate_add_to_cart_data(data):
-    """validate and return required data fields"""
-    product_id = verify_product_id(get_from_data(data, 'product_id'))
-    quantity = verify_quantity(get_from_data(data, 'quantity'))
-    return product_id, quantity
-
-def validate_delete_from_cart_data(data):
-    """validate and return required data fields"""
-    return verify_product_id(get_from_data(data, 'product_id'))
-
-def validate_update_item_data(data):
-    """validate and return required data fields"""
-    return verify_quantity(get_from_data(data, 'quantity'))
 
 # GET api/users/{user_id}/cart
 @app_views.route('/users/<string:user_id>/cart', methods=['GET'], strict_slashes=False)
@@ -75,23 +63,28 @@ def get_cart(user_id):
 def add_to_cart(user_id):
     """Add product to cart"""
     user = get_user_by_id(user_id)
-    cart = user.cart
-    cart_id = cart.id
+
     if not request.get_json():
         abort(400, description="Not a JSON")
     data = request.get_json()
-    (productId, quantity) = validate_add_to_cart_data(data)
+    productId = verify_product_id(get_from_data(data, 'product_id'))
+    quantity = verify_quantity(get_from_data(data, 'quantity'))
 
     # check if product already exists in cart
     foundInCart = False
+    cart = user.cart
     for item in cart.cart_items:
         if item.product_id == productId:
             # if item already in cart increase it's quantity
             foundInCart = True
             item.quantity += quantity
+            item.save()
             break
 
+    # If item does not exist in cart, create a new one 
+    # and associate it with cart
     if not foundInCart:
+        cart_id = cart.id
         cartItem = CartItem(cart_id=cart_id, **data)
         cartItem.save()
 
@@ -114,14 +107,14 @@ def delete_from_cart(user_id):
     """delete product from cart"""
 
     user = get_user_by_id(user_id)
-    cart = user.cart
 
     if not request.get_json():
         abort(400, description="Not a JSON")
     data = request.get_json()
-    productId = validate_delete_from_cart_data(data)
+    productId = verify_product_id(get_from_data(data, 'product_id'))
 
     foundInCart = False
+    cart = user.cart
     for item in cart.cart_items:
         if item.product_id == productId:
             foundInCart = True
@@ -149,18 +142,26 @@ def delete_from_cart(user_id):
 def update_in_cart(user_id, product_id):
     """update item of cart"""
     user = get_user_by_id(user_id)
-    cart = user.cart
+    productId = verify_product_id(product_id)
     if not request.get_json():
         abort(400, description="Not a JSON")
     data = request.get_json()
-    productId = validate_update_item_data(data)
+    quantity = verify_quantity(get_from_data(data, 'quantity'))
+
 
     foundInCart = False
+    cart = user.cart
     for item in cart.cart_items:
         if item.product_id == productId:
             foundInCart = True
-            # remove item from cart when found
-            item.delete()
+            # update quantity
+            if quantity == 0:
+                # if quantity = 0 delete item
+                item.delete()
+            else:
+                item.quantity = quantity
+                item.save()
+            # save changes
             cart.save()
             break
 
@@ -172,7 +173,7 @@ def update_in_cart(user_id, product_id):
         "cart_items": [item.dict_format() for item in cart.cart_items]
     }
     apiResponse = {
-        "message": f"Product with ID {productId} removed from cart successfully",
+        "message": f"Quantity of product with ID {productId} was upadated successfully",
         "cart": {**cart.dict_format(), **items}
     }
     return make_response(jsonify(apiResponse), 200)
